Move decorators to Module.prepend
Besides replacing class_eval with prepend, this change is adding
the SpreePriceBooks namespace to these decorators to ensure their
class name remains unique.
