+++ app/models/spree/spree_price_books/product_decorator.rb
@@ -1,6 +1,7 @@
-Spree::Product.class_eval do
-
-  has_many :price_books, through: :master
+module Spree::SpreePriceBooks::ProductDecorator
+  def self.prepended(base)
+    base.has_many :price_books, through: :master
+  end
 
   def master_price_for(price_book)
     Spree::Price.where(
@@ -13,5 +14,6 @@ Spree::Product.class_eval do
     price = master_price_for(price_book)
     Spree::Money.new(price.amount, currency: price_book.currency)
   end
-
 end
+
+Spree::Product.prepend Spree::SpreePriceBooks::ProductDecorator
